requirements.txt
pyudev>=0.24.1
zfs-autobackup>=3.2
PyYAML>=6.0.1